llm_service/requirements.txt
paho-mqtt
google-generativeai
langchain
langchain-google-genai
supabase